* tag 'LA.UM.8.4.r1-05700-8x98.0' of https://source.codeaurora.org/quic/la/kernel/msm-4.4:
  msm: adsprpc: Fix array index underflow problem
  profiler: Fix compilation errors
  fbdev: msm: fix ret value while translating to mdp format
  clk: msm: Fix compilation errors with clang
  ath10k: Fix compilation issue
  drm/msm/sde: Fix clang compilation in SDE driver
  ais: cam_soc_api: Fix compilation error
  msm: camera: reset atomic variable
  msm: v4l2loopback: Fix clang compilation error
  msm: Fix compilation error
  asoc: msm-8998: Add new FE's for TDM Primary
  diag: dci: Synchronize dci mempool buffers alloc and free
  headers_lib: rename Android.bp
  kbuild: skip install/check of headers right under uapi directories
  uapi: export all arch specifics directories
  kernel_headers: Fix headers not detected in incremental builds
  kernel headers: Update comparison logic
  Kbuild : Use no-export-headers to avoid exporting headers
  kernel_headers: kernel headers library package
  kbuild: replace genhdr-y with generated-y
  uapi: export all headers under uapi directories
  uapi: sound: move necessary sound headers under uapi
  kernel: Fix build errors with LLVM
  Revert "ARM: dts: msm: Remove cpu max frequency restriction on cold temp for msm8996."
  ARM: dts: msm: Remove cpu max frequency restriction on cold temp for msm8996.
  input: touchpanel: Fix compilation errors with clang
  asoc: msm-8998: TDM device grouping changes for capture path
  kbuild, x86: Track generated headers with generated-y
  kbuild: Remove stale asm-generic wrappers
  ARM: put types.h in uapi
  msm: ais: jpegdma: Fix uninitialized variable
  Makefile.headersinst: remove destination-y option
  msm: adsprpc: DSP device node to provide restricted access to ADSP/SLPI

Conflicts:
	.gitignore
	drivers/mmc/core/sdio_cis.c

+++ drivers/clk/msm/clock-osm.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2018, The Linux Foundation. All rights reserved.
+ * Copyright (c) 2016-2018, 2020, The Linux Foundation. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 and
@@ -272,7 +272,7 @@ static void __iomem *debug_base;
 #define ACD_REG_RELATIVE_ADDR_BITMASK(addr) \
 			(1 << (ACD_REG_RELATIVE_ADDR(addr)))
 
-#define FIXDIV(div) (div ? (2 * (div) - 1) : (0))
+#define FIXDIV(div) ((int)div ? (2 * (div) - 1) : (0))
 
 #define F(f, s, div, m, n) \
 	{ \
